requirements.txt
asgiref==3.8.1
attrs==25.4.0
Django==5.1.1
django-debug-toolbar==6.1.0
django-extensions==3.2.3
django-silk==5.4.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gprof2dot==2025.4.14
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==10.4.0
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.1
typing_extensions==4.15.0
tzdata==2024.2
uritemplate==4.2.0
